// src/enums/notification.enums.ts
export enum NotificationType {
    ORDER_CREATED = 'ORDER_CREATED',
    ORDER_ACCEPTED = 'ORDER_ACCEPTED',
    ORDER_REJECTED = 'ORDER_REJECTED',
    PAYMENT_CONFIRMED = 'PAYMENT_CONFIRMED',
    WITHDRAWAL_CONFIRMED = 'WITHDRAWAL_CONFIRMED',
    SHIPPED = 'SHIPPED',
    DELIVERED = 'DELIVERED',
    CHAT_MESSAGE = 'CHAT_MESSAGE'
}